requirements.txt
aiogram==3.10.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
mysql-connector-python==8.2.0
aiohttp==3.9.1
